greaterThanOrEqual tests and rename operator
