+++ src/askscript/__tests__/09-operators/operators-08-greater_than_or_equal.test.result.ts
@@ -1,0 +1,1 @@
+export const expectedResult = 'It works!';
