+++ src/askvm/resources/math/index.ts
@@ -3,7 +3,8 @@ import { minus } from './minus';
 import { sum } from './sum';
 import { times } from './times';
 import { lessThan } from './lessThan';
-import { greaterOrEq } from './greaterOrEq';
+import { greaterThan } from './greaterThan';
+import { greaterThanOrEqual } from './greaterThanOrEqual';
 
 Object.assign(exports, {
   '+': sum,
@@ -11,14 +12,16 @@ Object.assign(exports, {
   '*': times,
   '/': divideBy,
   '<': lessThan,
-  '>=': greaterOrEq,
+  '>': greaterThan,
+  '>=': greaterThanOrEqual,
 });
 
 export * from './ceil';
 export * from './floor';
 export * from './divideBy';
 export * from './lessThan';
-export * from './greaterOrEq';
+export * from './greaterThan';
+export * from './greaterThanOrEqual';
 export * from './max';
 export * from './minus';
 export { sum as plus, sum } from './sum';
